chore: fixed WAL file system, offset problems. Adding a flatbuf based metadata section. Currently unable to achive without zero copy
chore: major logical bug in header write.
Classic chicken-egg problem. Need to append the side of the stream in the header, however, the stream is lazy and not completely in memory. We now read the stream into memory (read INEFFECIENT).

chore: Fix offset errors in the python block validator

chore: fix the WAL fs - offset management, flatbuf for metadata